pyproject.toml
[tool.poetry]
name = "car-logos-dataset"
version = "0.1.0"
description = ""
authors = ["965311532 <gabriele.armento@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
tqdm = "^4.64.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
